EC-243 Rename package
