+++ s3-source-connector/src/integration-test/java/io/aiven/kafka/connect/s3/source/ConnectRunner.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.connect.source.s3;
+package io.aiven.kafka.connect.s3.source;
 
 import static org.assertj.core.api.Assertions.assertThat;
