+++ s3-source-connector/src/main/java/io/aiven/kafka/connect/s3/source/Version.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.connect.source.s3;
+package io.aiven.kafka.connect.s3.source;
 
 import java.io.InputStream;
 import java.util.Properties;
